chasm/lib/activity: split activity.go by concern
activity.go was 2225 lines and 81 methods on *Activity, mixing eight
independent concerns. Split it into files named for those concerns,
following chasm/lib/callback (the other single-component CHASM lib),
which uses bare file names rather than a component prefix.

This commit is a pure move: no declaration was edited, added or removed.

  activity.go            2225 -> 709   the component: state, CHASM
                                       contracts, construction, and the
                                       worker-facing entry points
  operator_commands.go        -> 641   update-options, cancel, pause,
                                       unpause, reset, deferred intents
  attempt.go                  -> 427   one attempt's lifecycle: when it
                                       dispatches, the timeout tasks that
                                       bound it, recording its failure,
                                       and whether another follows
  responses.go                -> 292   projection onto API protos
  metrics.go                  -> 196   metrics handlers and emitters
  outcome.go                  ->  38   recording the terminal outcome
  activity_tasks.go           -> tasks.go   (rename; no content change)

activity_test.go is split to match, since Go pairs foo.go with
foo_test.go. No metrics_test.go: the emitters are asserted inside the
behaviour tests, so a separate file would have held one borrowed test.

Two notes on where the attempt.go boundary falls, since it is the one
line here that needs explaining:

- The ScheduleToClose helpers (scheduleToCloseDeadline,
  reissueScheduleToClose) live in attempt.go even though ScheduleToClose
  is activity-scoped, not attempt-scoped - ActivityState.schedule_to_close_stamp
  exists separately from ActivityAttemptState.stamp precisely because it
  "spans the full activity lifetime". They are here because they are
  emitted alongside the attempt timers, and because ScheduleToClose is
  the only one of the four timeouts with named helpers at all: the other
  three compute their deadlines inline at each site. Giving the four
  timeouts a uniform shape is follow-up work, not a move.

- createAddActivityTaskRequest is in attempt.go next to
  newActivityDispatchTask rather than in activity.go, so that both
  builders on the dispatch path are in one place.

recordScheduleToStartOrCloseTimeoutFailure is in outcome.go rather than
with the other timeout code: it sets Outcome.Variant, which is what
outcome.go is for.

The "note on times and terminology" comment moves to attempt.go, because
it refers to "the comment below on nextAttemptDispatchTime", which moves
there. Two new files gain a one-line comment naming their scope.

Verified with .task/ast-diff, which emits a canonical position-independent
dump of the package AST at two revisions and diffs it: against the parent
commit, all 322 declarations plus their doc comments and the 72-entry
import union are BYTE-IDENTICAL. The only delta is added free-floating
comments, which are the file-scope comments above. The two renamed files
are byte-identical to their originals.

Not addressed here, deliberately: statemachine.go still carries logic
inline in 11 of its 18 transition bodies while the other 7 delegate to
named methods. Making it declarative is a refactor, not a move, so it is
the next commit.

Run the tests from the repo root with:
  export TEMPORAL_TEST_LOG_LEVEL=ERROR TEMPORAL_TEST_LOG_STACKTRACE_LEVEL=off
  go test ./chasm/lib/activity/...
